clients/python/schema_registry/requirements.txt
avro==1.10.2
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
kafka-python==2.0.2
requests==2.26.0
urllib3==1.26.7
